move Transaction, ErgoBoxCandidate and other chain types to sigma-tree crate;
