+++ sigma-tree/Cargo.toml
@@ -9,6 +9,8 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 sigma-ser = { path = "../sigma-ser" }
+indexmap = "1.3.2"
+base16 = "0.2.1"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.10"
