+++ sigma-tree/src/chain/box_id.rs
