+++ sigma-tree/src/constants.rs
