+++ sigma-tree/src/lib.rs
@@ -11,10 +11,11 @@
 #![deny(missing_docs)]
 
 mod ast;
+mod constants;
 mod data;
-mod ergo_tree;
 mod eval;
 mod serialization;
 mod types;
 
-pub use ergo_tree::*;
+pub mod chain;
+pub mod ergo_tree;
